pyproject.toml
[project]
name = "peakguard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "commitizen>=4.13.9",
    "pre-commit>=4.5.1",
    "pyyaml>=6.0",
    "requests>=2.31",
    "yfinance>=0.2.40",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/peakguard"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.14",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
